app/layers/database/requirements.txt
mysql-connector-python==8.0.29
protobuf==3.20.1